requirements.txt
asgiref==3.7.2
Django==5.0
django-ckeditor==6.7.1
django-environ==0.11.2
django-js-asset==2.2.0
django-shortcuts==1.6
pillow==10.2.0
sqlparse==0.4.4
tzdata==2023.4
